Feature main movies dashboard (#3)
* Add movies data and interfaces

Added the movies data and interfaces into the project.

* Add movies json into the app

Added the movies json data into the app and displayed it in a table.

* Add table pagination

Added movie table pagination and removed unnecessary interfaces.

* Add material table and logic

Added material table to the project and some logic for the dashboard.

* Add dashboard and navbar styles

Added dashboard and navbar styles with the help of material ui

* Add styles for login component

Added styles for the login component, also included the option to login anonymously.

* Update onchange auth method

Updated the onauthstatechange method to remove console logs. Also deleted App component which wasn't being used.

* Remove unused filter

Removed unused custom filter for movie genres.